hotfix: fix not good calculate

+++ internal/render/render.go
@@ -13,8 +13,8 @@ import (
 func CliRender(layer models.Layer, z, x, y int) (image.Image, error) {
 	coef := math.Pow(2, float64(z))
 	maxSize := min(layer.Width, layer.Height)
-	readSize := int(float64(maxSize) / float64(coef))
-	if x*readSize >= layer.Width || y*readSize >= layer.Height {
+	readSize := float64(maxSize) / float64(coef)
+	if float64(x)*readSize >= float64(layer.Width) || float64(y)*readSize >= float64(layer.Height) {
 		return nil, errors.New("Выход за границы")
 	}
 	tmpFile, err := os.CreateTemp("../resource", "tile_*.png")
@@ -26,10 +26,10 @@ func CliRender(layer models.Layer, z, x, y int) (image.Image, error) {
 	defer os.Remove(tmpPath)
 	defer os.Remove(tmpPath + ".aux.xml")
 	cmd := exec.Command("gdal_translate", "-srcwin",
-		fmt.Sprintf("%d", x*readSize),
-		fmt.Sprintf("%d", y*readSize),
-		fmt.Sprintf("%d", readSize),
-		fmt.Sprintf("%d", readSize),
+		fmt.Sprintf("%d", int(float64(x)*readSize)),
+		fmt.Sprintf("%d", int(float64(y)*readSize)),
+		fmt.Sprintf("%d", int(readSize)),
+		fmt.Sprintf("%d", int(readSize)),
 		"-outsize",
 		fmt.Sprintf("%d", layer.TileSize),
 		fmt.Sprintf("%d", layer.TileSize),
